+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 1.4.0 / 2021-05-27
 
 * Add config collector, enabled by default